requirements.txt
autopep8==1.6.0
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.19.2
SQLAlchemy==1.4.31
toml==0.10.2
Werkzeug==2.0.3
